+++ addon/nodes/paragraph.ts
@@ -19,6 +19,7 @@ export const paragraph: NodeSpec = {
       default: 0,
     },
   },
+  indentable: true,
   // defining: true,
   parseDOM: [
     {